Reorganized main balrog code into many separate files. First step towards #47. Closees #38.
