+++ balrog/__init__.py
@@ -1,6 +1,10 @@
-from . import prep
-from . import sim
-from . import files
+from . import config
+from . import grid
+from . import tile
+from . import chip
+from . import fileio
+from . import mathutil
+from . import balobject
 from . import ngmix_catalog
 from . import des_star_catalog
 from . import meds_catalog
