+++ balrog/balobject.py
@@ -1,0 +1,25 @@
+#-------------------------------------------------------------------------------
+# DESCRIPTION
+
+class BalObject(object):
+    '''
+    '''
+
+    def __init__(self):
+        pass
+
+class Galaxy(object):
+    '''
+    #TODO: Do we need a galaxy class? (probably not)
+    '''
+
+    def __init__(self):
+        pass
+
+class Star(object):
+    '''
+    #TODO: Do we need a star class? (probably not)
+    '''
+
+    def __init__(self):
+        pass
